requirements.txt
requests
pillow
imagehash
python-dotenv
